+++ CHANGELOG.md
@@ -5,6 +5,7 @@ This is the changelog for [AWS Architect](readme.md).
 * Add support for creating dynamoDB Tables.
 * Separate policy examples into their own documents.
 * Allow specifying specific port to run on.
+* Deploy content directory to s3.
 
 ## 2.0 ##
 * All configuration is now contained in the api files. Generation, run, and testing is completed via index.js.  New templates created.